feat: added prompt for Gemini 1.5
